feat(espectro): o fantasma não toca o chão, e atravessa — fatia 2
TRAÇO, e não postura. Postura é decisão do turno e some no fim do slot;
guardar "incorpóreo" ali faria o fantasma virar corpo toda vez que o slot
acabasse — ele seria fantasma só no primeiro terço de cada turno. O traço
vem do ELEMENTO e vale sempre, e há teste disso.

"Fora do chão" virou UMA pergunta (`IsOffTheGround`). Três fases faziam
essa conta — o dano da casa, o escorregão e o atolo — e voar/submergir já
respondiam a duas delas por caminhos separados. Três cópias concordariam
até a primeira edição, e o fantasma foi quem forçou a unificação: ele é a
primeira criatura que está fora do chão sem ter feito nada.

Disso decorre tudo de uma vez, sem uma exceção para cada terreno: a brasa
não o queima, o gelo não o faz escorregar, a lama não o atola. Não é
imunidade concedida — é consequência de não encostar.

ATRAVESSAR é ação, não permanência. Passar por dentro do tronco de graça
e para sempre faria a casa bloqueada deixar de existir para ele, e o
tabuleiro perderia metade do sentido contra um tipo. Gastando a ação, ele
passa — e o tronco CONTINUA em pé, porque isto não é derrubar: quem
derruba abre a passagem para todo mundo, e é outra jogada, com outro
custo e outro evento. Tem teste dos dois.

A recusa de quem tem corpo entra pela porta que a fundura abriu — falha
ALTA com evento próprio, e não ação sumindo calada.

PostureFlags virou uint16. Eu tinha escrito, na fatia da lama, que
`Slowed` era o último bit e que o próximo obrigaria a alargar. Obrigou.
Muda o hash de todo estado, e é deliberado: a alternativa era espremer
duas ideias no mesmo bit.

Falta: o físico não acertar e a resposta da luz (fatia 3), drenar
(fatia 4) e os pets (fatia 5).

Bateria: 536/536, cinco sondas limpas.

+++ Source/BattleSim/Private/Battle/BattlePhaseResolution.cpp
@@ -123,18 +123,18 @@ void BattlePhases::ApplyResolution(
 		// novo entraria no hash do estado e invalidaria os snapshots de
 		// determinismo de cenários que nem usam estas ações.
 		const bool bEstavaCamuflado =
-			(Pet.PostureFlags & static_cast<uint8>(EBattlePostureFlags::Camouflaged)) != 0;
+			(Pet.PostureFlags & static_cast<uint16>(EBattlePostureFlags::Camouflaged)) != 0;
 		const bool bEstavaSubmerso =
-			(Pet.PostureFlags & static_cast<uint8>(EBattlePostureFlags::Underground)) != 0;
+			(Pet.PostureFlags & static_cast<uint16>(EBattlePostureFlags::Underground)) != 0;
 
 		Pet.PostureFlags = 0;
 		if (bEstavaCamuflado)
 		{
-			Pet.PostureFlags = static_cast<uint8>(EBattlePostureFlags::Revealing);
+			Pet.PostureFlags = static_cast<uint16>(EBattlePostureFlags::Revealing);
 		}
 		else if (bEstavaSubmerso)
 		{
-			Pet.PostureFlags = static_cast<uint8>(EBattlePostureFlags::Emerging);
+			Pet.PostureFlags = static_cast<uint16>(EBattlePostureFlags::Emerging);
 		}
 	}
